anv: Add support for batch buffers in anv_slab_bo in Xe KMD
Because of the ANV_BO_ALLOC_CAPTURE flag, batch buffers were not
allowed to be memory-pooled yet.
So to workaround that here adding a new anv_bo_slab_heap heap just
for batch buffers.

For now that will only work in Xe KMD as i915 requires more changes
to support it.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>