brw/brw_lex: jump label targets need to be at the end of the line
The Gen assembly grammar defined by Mesa is a little different than
the one defined by BSpec in pages "Instruction Syntax" (56717) and
"Grammar" (56736). One of the differences is that in Mesa, the
presence of the ':' character before types is optional (e.g., we
accept both g16<1>:F and g16<1>F), while for BSpec it's mandated.
Well, it is optional, until an ambiguity in our rules prevents us from
using it in send instructions.

In my attempt to try to make Mesa print ':' before the register types
(see the next commit for an explanation on why), I noticed that for
send instructions specifically, our parser was failing to parse
something like:

  send(1)         g9:UD           g47:UD          null:UD         0x6210c500                0x03000000

Due to the Flex's longest match rule, g9:UD (and g47:UD) would match
JUMP_LABEL_TARGET with g9: instead of matching GENREG with g9 and then
TYPE_UD with :UD. BSpec's page "Grammar" (56736) says that a label
definition should end with a new line (i.e., no other tokens can come
after the label definition until the end of the line), so in this
patch we change the JUMP_LABEL_TARGET rule to allow only white space
and tabs before the end of the line. This makes it stop matching
g9:UD, so we get the desired GENREG followed by TYPE_UD.

Notice that these types we print for the send instruction don't really
make a whole lot of sense, and they are not how BSpec defines sends,
but changing that would be way more complex, and there's no need for
it. I just need us to consistently optionally accept the ':'
character for now.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>