anv: Remove check for debug_marker_count
Remove checks for debug_marker_counter == 0 when building
the bvh. Seems like we turn this off when tracking each step
within the runtime code for bvh building, but we miss out on
useful compute data. With this, Perfetto traces aligns with
other profiling tools in showing the build steps + the dispatches.

Signed-off-by: Michael Cheng <michael.cheng@intel.com>