lima: ppir: fix usage of comp->var_nodes for SSAs
Currently var_nodes array size is x4 of nir_defs, because we need to
track writes for each individual component of a register. And while
registers are indexed in this array correctly (i.e. reg->index << 2),
SSAs are not, as a result, reg #4 will share a slot in var_nodes with
SSA #16.

Fix that by properly shifting SSA index

While we are at here, track write after read deps properly as well by
checking individual components.

Signed-off-by: Vasily Khoruzhick <anarsoul@gmail.com>