Meson: Rust: Bump minimum Meson version
Meson 1.7.0 is required to ensure that generated bindings
are suitable for use with the selected Rust compiler.

Meson doesn't pass `rust_std` to `rust.bindgen()`, even in 1.7.0,
however it does pass the Rust version through to bindgen which
ensures that generated bindings do not use e.g. `unsafe extern`
if the compiler does not support that.

Bindgen is bumped to 0.71.0 which has more permissive handling of
`--rust-target`, enabling (almost) any valid Rust version to be
passed as a target; -nightly and -beta* suffixes are an exception
to this, which an upstream ticket has been raised to address.

This behaviour could be improved further if later versions of Meson
pass an appropriate `--rust-edition` to bindgen (matching the `rust_std`
project option), which is supported from 0.71.0.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/12680
Bug: https://bugs.gentoo.org/946913
Signed-off-by: Matt Jolly <kangie@gentoo.org>