gitlab-ci: update alpine build script to use install-meson.sh
This is required because we need Meson 1.7.0 which is not currently in
the Alpine repos. Using this script requires that we explicitly install
a ninja impl; samurai is preferred on Alpine, and the `ninja-build`
package does not place ninja in PATH.

Additionally dependencies have been sorted as it was half-done.